refactor: move tests outside from src
