+++ tsconfig.json
@@ -27,6 +27,6 @@
     "baseUrl": "."
   },
 
-  "include": ["./src"],
+  "include": ["./src", "tests"],
   "exclude": ["node_modules"]
 }
